+++ test/source.test.js
@@ -18,6 +18,7 @@ var tmp = tm.join(require('os').tmpdir(),'mapbox-studio');
 var UPDATE = !!process.env.UPDATE;
 var server;
 
+var defaultsource = 'tmsource://' + tm.join(path.dirname(require.resolve('mapbox-studio-default-source')));
 var localsource = 'tmsource://' + path.join(__dirname,'fixtures-localsource');
 var tmppath = tm.join(tmp, 'Source ШЖФ - ' + +new Date);
 
@@ -197,9 +198,17 @@ test('remote: error bad protocol', function(t) {
     });
 });
 
-test('remote: noop remote write', function(t) {
-    source.save({id:'mapbox:///mapbox.mapbox-streets-v2'}, function(err, source) {
+test('remote: noop refresh', function(t) {
+    source.refresh({id:'mapbox:///mapbox.mapbox-streets-v2'}, function(err, source) {
         t.ifError(err);
+        t.end();
+    });
+});
+
+test('remote: save error', function(t) {
+    source.save({id:'mapbox:///mapbox.mapbox-streets-v2'}, function(err, source) {
+        t.ok(err);
+        t.ok(/^Error: Cannot save remote source/.test(err.toString()));
         t.end();
     });
 });
@@ -242,29 +251,47 @@ test('local: loads via tilelive', function(t) {
     });
 });
 
-test('local: saves source in memory', function(t) {
-    testutil.createTmpProject('source-save', localsource, function(err, tmpid, info) {
-    assert.ifError(err);
-
-    source.save(_({id:source.tmpid()}).defaults(info), function(err, source) {
-        t.ifError(err);
-        t.ok(source);
-        t.end();
-    });
-
-    });
-});
-
-test('local: saves source (invalid)', function(t) {
+test('local: refresh source in memory', function(t) {
     testutil.createTmpProject('source-save', localsource, function(err, tmpid, info) {
         assert.ifError(err);
-        source.save(_({id:source.tmpid(), minzoom:-1}).defaults(info), function(err, source) {
-            assert.equal(err.toString(), 'Error: minzoom must be an integer between 0 and 22', 'source.save() errors on invalid style');
+        source.refresh(_({id:source.tmpid()}).defaults(info), function(err, source) {
+            t.ifError(err);
+            t.ok(source);
             t.end();
         });
     });
 });
 
+test('local: refresh source (invalid minzoom)', function(t) {
+    testutil.createTmpProject('source-save', defaultsource, function(err, tmpid, info) {
+        assert.ifError(err);
+        source.refresh(_({id:source.tmpid(), minzoom:-1}).defaults(info), function(err, source) {
+            assert.equal(err.toString(), 'Error: minzoom must be an integer between 0 and 22', 'source.refresh() errors on invalid source');
+            t.end();
+        });
+    });
+});
+
+test('local: refresh source (invalid center)', function(t) {
+    testutil.createTmpProject('source-save', defaultsource, function(err, tmpid, info) {
+        assert.ifError(err);
+        source.refresh(_({id:source.tmpid(), minzoom:3, center:[0,0,0]}).defaults(info), function(err, source) {
+            assert.equal(err.toString(), 'Error: center zoom value must be greater than or equal to minzoom 3', 'source.refresh() errors on invalid source');
+            t.end();
+        });
+    });
+});
+
+test('local: save temporary errors', function(t) {
+    testutil.createTmpProject('source-save', localsource, function(err, tmpid, info) {
+        assert.ifError(err);
+        source.save(_({id:source.tmpid()}).defaults(info), function(err, source) {
+            t.ok(err);
+            t.ok(/^Error: Cannot save temporary source/.test(err.toString()));
+            t.end();
+        });
+    });
+});
 
 test('local: saves source to disk', function(t) {
     testutil.createTmpProject('source-save', localsource, function(err, tmpid, data) {
@@ -304,6 +331,23 @@ test('local: saves source to disk', function(t) {
         }, 1000);
     });
 
+    });
+});
+
+test('local: save as tmp => perm', function(t) {
+    var tmpid = 'tmpsource://' + tm.join(__dirname, '/fixtures-localsource');
+    source.info(tmpid, function(err, data) {
+        t.ifError(err);
+        var permid = path.join(tmp, 'source-saveas-' + Math.random().toString(36).split('.').pop());
+        source.save(_({_tmp:tmpid, id:permid}).defaults(data), function(err, source) {
+            t.ifError(err);
+            t.ok(source);
+            t.equal(source.data._tmp, false);
+            var tmpdir = tm.parse(permid).dirname;
+            t.ok(fs.existsSync(tm.join(tmpdir,'data.yml')));
+            t.ok(fs.existsSync(tm.join(tmpdir,'data.xml')));
+            t.end();
+        });
     });
 });